userbot/requirements.txt
telethon
aiosqlite
python-dotenv